+++ packages/content/src/editing/codegen/preview.test.ts
@@ -2,6 +2,7 @@
 /* eslint-disable no-unused-expressions */
 import { expect } from 'chai';
 import sinon from 'sinon';
+import { constants } from '@sitecore-content-sdk/core';
 import {
   getDesignLibraryComponentPropsEvent,
   getDesignLibraryImportMapEvent,
@@ -10,7 +11,7 @@ import {
   ImportEntry,
   ComponentImport,
   buildComponentDependencies,
-  getDesignLibraryComponentPreviewErrorEvent,
+  getDesignLibraryErrorEvent,
   DesignLibraryPreviewError,
   ComponentPreviewEventArgs,
   addStyleElement,
@@ -22,6 +23,8 @@ import {
   GeneratedComponentData,
 } from './preview';
 import { NativeDataFetcher } from '@sitecore-content-sdk/core';
+
+const { ERROR_MESSAGES } = constants;
 
 describe('design library codegen', () => {
   let debugSpy: sinon.SinonSpy;
@@ -321,7 +324,9 @@ describe('design library codegen', () => {
           .to.be.true;
 
         const errorLogArgs = errorSpy.getCall(0).args;
-        expect(errorLogArgs[0]).to.equal('Component Library: sending error event');
+        expect(errorLogArgs[0]).to.equal(
+          `Component Library: sending error event. ${ERROR_MESSAGES.CONTACT_SUPPORT}`
+        );
 
         const errorEvent = errorLogArgs[1];
 
@@ -357,7 +362,9 @@ describe('design library codegen', () => {
           .to.be.true;
 
         const errorLogArgs = errorSpy.getCall(0).args;
-        expect(errorLogArgs[0]).to.equal('Component Library: sending error event');
+        expect(errorLogArgs[0]).to.equal(
+          `Component Library: sending error event. ${ERROR_MESSAGES.CONTACT_SUPPORT}`
+        );
 
         const errorEvent = errorLogArgs[1];
 
@@ -766,9 +773,9 @@ describe('design library codegen', () => {
     });
   });
 
-  describe('getDesignLibraryComponentPreviewErrorEvent', () => {
+  describe('getDesignLibraryErrorEvent', () => {
     it('should return a valid component preview "render" error event', () => {
-      const errorEvent = getDesignLibraryComponentPreviewErrorEvent(
+      const errorEvent = getDesignLibraryErrorEvent(
         'uid-1',
         'custom-error',
         DesignLibraryPreviewError.Render
@@ -780,7 +787,7 @@ describe('design library codegen', () => {
     });
 
     it('should return a valid component preview "render-init" error event', () => {
-      const errorEvent = getDesignLibraryComponentPreviewErrorEvent(
+      const errorEvent = getDesignLibraryErrorEvent(
         'uid-1',
         'custom-error',
         DesignLibraryPreviewError.RenderInit
@@ -792,6 +799,56 @@ describe('design library codegen', () => {
           uid: 'uid-1',
           error: 'custom-error',
           type: DesignLibraryPreviewError.RenderInit,
+        },
+      });
+    });
+
+    it('should return valid load import map error event', () => {
+      const errorEvent = getDesignLibraryErrorEvent(
+        'uid-1',
+        'error during loading import map',
+        DesignLibraryPreviewError.ImportMapLoad
+      );
+
+      expect(errorEvent).to.deep.equal({
+        name: 'component:generation:error',
+        message: {
+          uid: 'uid-1',
+          error: 'error during loading import map',
+          type: DesignLibraryPreviewError.ImportMapLoad,
+        },
+      });
+    });
+
+    it('should return valid import map missing error event', () => {
+      const errorEvent = getDesignLibraryErrorEvent(
+        'uid-1',
+        'import map is missing',
+        DesignLibraryPreviewError.ImportMapMissing
+      );
+      expect(errorEvent).to.deep.equal({
+        name: 'component:generation:error',
+        message: {
+          uid: 'uid-1',
+          error: 'import map is missing',
+          type: DesignLibraryPreviewError.ImportMapMissing,
+        },
+      });
+    });
+
+    it('should return valid fetch error event', () => {
+      const errorEvent = getDesignLibraryErrorEvent(
+        'uid-1',
+        'error during fetching component data',
+        DesignLibraryPreviewError.GeneratedComponentFetch
+      );
+
+      expect(errorEvent).to.deep.equal({
+        name: 'component:generation:error',
+        message: {
+          uid: 'uid-1',
+          error: 'error during fetching component data',
+          type: DesignLibraryPreviewError.GeneratedComponentFetch,
         },
       });
     });